Add files via upload
A support ticket is a request for assistance or support from a customer or user of a product or service. When a customer encounters an issue, has a question, or needs help with a product or service, they can submit a support ticket to the provider or vendor.

Submission: The customer submits a support ticket through various channels such as email, web forms, or support portals. They usually provide details about their issue or inquiry.

Ticket Creation: Upon submission, a support ticket is created in the provider's ticketing system. This ticket includes information such as the customer's contact details, the nature of the issue, and any relevant attachments.

Assignment: The ticket is assigned to a support agent or team member responsible for handling such requests.

Resolution: The support agent investigates the issue and works to resolve it. This may involve troubleshooting, providing guidance, or escalating the ticket to a higher level of support.

Communication: Throughout the process, there is often communication between the support agent and the customer to gather more information, provide updates on progress, and ultimately inform the customer of the resolution.

Closure: Once the issue is resolved or the customer's inquiry is adequately addressed, the support ticket is marked as closed.

Support tickets help streamline the support process by organizing and prioritizing customer inquiries, ensuring that each request receives proper attention and resolution. They also serve as a record of customer interactions, allowing organizations to track trends, analyze common issues, and improve their products or services over time.

Here is more info https://www.phpcrm.com/customer-help-desk-system/

You can download CRM software here https://www.phpcrm.com/download/